Fix overlap margin calculation.
The current code is only called when a new compressed byte is read,
not after the corresponding decompressed bytes are written. This
means that the contents produced by the last byte of input are not
taken into account, which can make a huge difference for instance on
files with long sequences of zeros at the end.